Housecleaning Update
So called because I updated many parts of the movement/player control script to be more refined. Changes include
-Updated move script
- Updated reflection/attack code

the reflection/attack code is not done yet. I will need to create the enemy ai with its projectile. THEN i can make the reflection work. Itll take a while, but using the same tutorials as before, it should be fine.