Reflected Projectile + a problem
I finally got the projectile reflection to work! The best part was that I didn't even need to use much tutorials. Just what I had in my code. 
Note to self: When I need to edit something in the code to allow another function, I need to simply change the code itself rather than watch a tutorial that has different code and try to add it on there. Hopefully it makes more sense in context?
A problem: Player still takes damage even after reflecting. Perhaps this is because the shield is a child of the player?